feat: place arbitrary day table-based pages side-by-side
Refactored the makefile to include a preprocessing step as described in
issue #2. The booklet.adoc was refactored to provide an interface to the
build system. It uses attributes which allows the build system to tell
booklet.adoc where to find the processed days table-pages without
passing the paths directly. This interface provides a level of
encapsulation and separation of concerns.

Closes #2.