exp: make schedule blank
Remove all planning from schedule such that each day is an empty table
of 5-minute increments.

This experiment is testing a less-rigid schedule structure in which I
can plan my time closer to when the time actually passes. I want to plan
my day based on achievable goals I can accomplish rather than a
category. My hypothesis is that SMART goals are a better way to stay
motivated and accomplish things versus just stating that your time is
for a certain period is being allocated to a specific responsibility.